make performance test output better

+++ ansible/octo-weasel/run-performance-test.sh
@@ -5,7 +5,13 @@ echo "$(date '+%Y-%m-%d %H-%M-%S') Starting ..."
 
 function execute_tests {
 
-  rm /var/log/apache2/access.log /var/log/mysql/mysql.log /var/log/mysql/mysql-slow.log
+  if [ -f /var/log/mysql/mysql.log ]; then
+    rm /var/log/mysql/mysql.log
+  fi
+  if [ -f /var/log/mysql/mysql-slow.log ]; then
+    rm /var/log/mysql/mysql-slow.log
+  fi
+  echo -n "" > /var/log/apache2/access.log
 
   echo "$(date '+%Y-%m-%d %H-%M-%S') Re-setup MySQL ..."
   mysql -e "DROP DATABASE owncloud; CREATE DATABASE owncloud; SET GLOBAL general_log = $1;"
@@ -18,10 +24,15 @@ function execute_tests {
   currentTime=$(date +%Y-%m-%d.%H-%M-%S)
   echo "$(date '+%Y-%m-%d %H-%M-%S') Running performance test ..."
   DAV_USER=admin DAV_PASS=admin /root/administration/performance-tests-c++/webdav-benchmark http://localhost/remote.php/webdav/ -csv > /tmp/performance-tests/$currentTime.csv
-  mv /var/log/mysql/mysql.log /tmp/performance-tests/mysql-general-query-$currentTime.log
-  mv /var/log/mysql/mysql-slow.log /tmp/performance-tests/mysql-slow-query-$currentTime.log
 
-  mv /var/log/apache2/access.log /tmp/performance-tests/access-$currentTime.log
+  if [ -f /var/log/mysql/mysql.log ]; then
+    mv /var/log/mysql/mysql.log /tmp/performance-tests/mysql-general-query-$currentTime.log
+  fi
+  if [ -f /var/log/mysql/mysql-slow.log ]; then
+    mv /var/log/mysql/mysql-slow.log /tmp/performance-tests/mysql-slow-query-$currentTime.log
+  fi
+
+  cp /var/log/apache2/access.log /tmp/performance-tests/access-$currentTime.log
 }
 
 echo "$(date '+%Y-%m-%d %H-%M-%S') Running WITHOUT general query logger ... "
